* [stringification] did some work to get stringification to work better * [with_sl] fixed constructor, now takes any type, not only directly T * [named] fixed stringification of `named` * [source_location] `source_location` is now stringifiable * [geometry] some better names and additional preliminary functions for iterations * [geometry] added some additional functions for rec2 to interop with geometry submodule * [tests] did some work to make the ghassanpl::tests system better; still VERY far from production code * small fixes

+++ include/ghassanpl/random.h
@@ -143,8 +143,10 @@ namespace ghassanpl::random
 		}
 	}
 
+	/// TODO: Should we move the below to random_seq?
+	
 	template <std::floating_point T = float>
-	constexpr T halton_sequence(unsigned index, unsigned base = 2)
+	constexpr T halton_sequence(size_t index, size_t base = 2)
 	{
 		auto result = T(0);
 		auto fraction = T(1);
